requirements.txt
groq
python-dotenv
rich
requests
fastapi
pydantic
uvicorn[standard]
websockets
